Fix the minor issue with compile & handle the out of range choice

+++ matrix.hpp
@@ -28,7 +28,7 @@ public://Define public method
     Matrix operator*(const Matrix &rhs) const;
     void set_value(std::size_t i, std::size_t j, int n);
     int get_value(std::size_t i, std::size_t j) const;
-    int get_size() const;
+    size_t get_size() const;
     int sum_diagonal_major() const;
     int sum_diagonal_minor() const;
     void swap_rows(std::size_t r1, std::size_t r2);
